+++ src/main/java/com/hrznstudio/galacticraft/block/entity/ElectricCompressorBlockEntity.java
@@ -176,7 +176,7 @@ public class ElectricCompressorBlockEntity extends ConfigurableMachineBlockEntit
     }
 
     @Override
-    public double getBaseEnergyConsumption() {
+    public int getBaseEnergyConsumption() {
         return Galacticraft.configManager.get().electricCompressorEnergyConsumptionRate();
     }
